:seedling: (FE:mock) サンプルフレンドの追加
