+++ frontend/pong/js/mocks/handlers/index.js
@@ -1,3 +1,4 @@
+import { handlers as friendsHandlers } from "./friends";
 import { handlers as healthHandlers } from "./health";
 import { handlers as usersHandlers } from "./users";
 import { handlers as webSocketHandlers } from "./webSocket";
@@ -5,5 +6,6 @@ import { handlers as webSocketHandlers } from "./webSocket";
 export const handlers = [
   ...healthHandlers,
   ...usersHandlers,
+  ...friendsHandlers,
   ...webSocketHandlers,
 ];
